[bugfix/feature] Fix get_synergy_table to take any order of inhibitors
Fix a bug in get_synergy_table that removed inhibitor B if both A+B and
B+A were present in the "Treatment" column.
The fix fixes the more general problem of deciding which drug is row or
column and provides a new argument to force this behavious.
Add documentation for the new arguments and update the documentation.